replaced int with Int64
ToDo: replace AccountTask id with int64

+++ src/SuperSmart/SuperSmart.Core/Data/ViewModels/CreateTeachingClassViewModel.cs
@@ -9,8 +9,8 @@ namespace SuperSmart.Core.Data.ViewModels
         [MinLength(1, ErrorMessage = "Please enter a designation")]
         public string Designation { get; set; }
         [Required(ErrorMessage = "Please enter a number of education years")]
-        [Range(1, int.MaxValue, ErrorMessage = "Please enter a number of education years")]
-        public int NumberOfEducationYears { get; set; }
+        [Range(1, Int64.MaxValue, ErrorMessage = "Please enter a number of education years")]
+        public Int64 NumberOfEducationYears { get; set; }
         [Required(ErrorMessage = "Please enter the time when the education started")]
         public DateTime Started { get; set; }
     }
