+++ src/SuperSmart/SuperSmart.Core/Persistence/Implementation/TeachingClassPersistence.cs
@@ -216,7 +216,7 @@ namespace SuperSmart.Core.Persistence.Implementation
         /// </summary>
         /// <param name="id"></param>
         /// <param name="loginToken"></param>
-        public void Delete(int id, string loginToken)
+        public void Delete(Int64 id, string loginToken)
         {
             Guard.NotNullOrEmpty(loginToken);
 
@@ -252,7 +252,7 @@ namespace SuperSmart.Core.Persistence.Implementation
         /// </summary>
         /// <param name="id">teaching class id</param>
         /// <param name="loginToken"></param>
-        public void ChangeReferral(int id, string loginToken)
+        public void ChangeReferral(Int64 id, string loginToken)
         {
             Guard.NotNullOrEmpty(loginToken);
